Add the rest of files from `octopress new` dir
As per the previous commit message, the rest of these files are from
the test `octopressTest` directory on my local machine. That directory
was created by running `octopress new <PATH>` (`<PATH>` being
`octopressTest` in this case). Since there was already an `index.html`
in this repo I renamed that one to `_index.html`.