Add old index.html as a post, prettify JSON
The old `index.html` is essentially the blog post “Things that have
tripped me up while writing JavaScript” that my old GitHub pages blog
was solely comprised of. I turned it into a Jekyll-formatted post and
am uploading it as such in this commit.

The Markdown content of that old post was contained in `params.json`,
but it came up as one huge blob of a JSON file so I installed and used
the Pretty JSON package in Sublime Text to prettify it. That’s what’s
being committed.